# [FIX] point_of_sale: speed-up picking creation at closing