--- a/addons/point_of_sale/tests/test_point_of_sale_flow.py
+++ b/addons/point_of_sale/tests/test_point_of_sale_flow.py
@@ -406,13 +406,19 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
         self.pos_config.current_session_id.action_pos_session_closing_control()
 
     def test_order_to_picking02(self):
-        """ This test is similar to test_order_to_picking except that this time, there are two products:
-            - One tracked by lot
+        """ This test is similar to test_order_to_picking except that this time, there are three products:
+            - One tracked by lot, with preexisting lot
+            - One tracked by lot, without preexisting lot
             - One untracked
-            - Both are in a sublocation of the main warehouse
+            - All are in a sublocation of the main warehouse
         """
-        tracked_product, untracked_product = self.env['product.product'].create([{
+        tracked_product_w_lot, tracked_product_wo_lot, untracked_product = self.env['product.product'].create([{
             'name': 'SuperProduct Tracked',
+            'type': 'product',
+            'tracking': 'lot',
+            'available_in_pos': True,
+        }, {
+            'name': 'SuperProduct Tracked No Lot',
             'type': 'product',
             'tracking': 'lot',
             'available_in_pos': True,
@@ -429,17 +435,18 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
         })
         lot = self.env['stock.production.lot'].create({
             'name': 'SuperLot',
-            'product_id': tracked_product.id,
+            'product_id': tracked_product_w_lot.id,
             'company_id': self.env.company.id,
         })
         qty = 2
-        self.env['stock.quant']._update_available_quantity(tracked_product, shelf1_location, qty, lot_id=lot)
+        self.env['stock.quant']._update_available_quantity(tracked_product_w_lot, shelf1_location, qty, lot_id=lot)
+        self.env['stock.quant']._update_available_quantity(tracked_product_wo_lot, shelf1_location, qty)
         self.env['stock.quant']._update_available_quantity(untracked_product, shelf1_location, qty)
 
         self.pos_config.open_session_cb()
         self.pos_config.current_session_id.update_stock_at_closing = False
 
-        untax, atax = self.compute_tax(tracked_product, 1.15, 1)
+        untax, atax = self.compute_tax(tracked_product_w_lot, 1.15, 1)
 
         for i in range(qty):
             pos_order = self.PosOrder.create({
@@ -449,11 +456,11 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
                 'partner_id': self.partner1.id,
                 'lines': [(0, 0, {
                     'name': "OL/0001",
-                    'product_id': tracked_product.id,
+                    'product_id': tracked_product_w_lot.id,
                     'price_unit': untax + atax,
                     'discount': 0.0,
                     'qty': 1.0,
-                    'tax_ids': [(6, 0, tracked_product.taxes_id.ids)],
+                    'tax_ids': [(6, 0, tracked_product_w_lot.taxes_id.ids)],
                     'price_subtotal': untax,
                     'price_subtotal_incl': untax + atax,
                     'pack_lot_ids': [[0, 0, {'lot_name': lot.name}]],
@@ -466,9 +473,19 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
                     'tax_ids': [(6, 0, untracked_product.taxes_id.ids)],
                     'price_subtotal': untax,
                     'price_subtotal_incl': untax + atax,
+                }), (0, 0, {
+                    'name': "OL/0003",
+                    'product_id': tracked_product_wo_lot.id,
+                    'price_unit': untax + atax,
+                    'discount': 0.0,
+                    'qty': 1.0,
+                    'tax_ids': [(6, 0, tracked_product_wo_lot.taxes_id.ids)],
+                    'price_subtotal': untax,
+                    'price_subtotal_incl': untax + atax,
+                    'pack_lot_ids': [[0, 0, {'lot_name': 'New SuperLot'}]],
                 })],
-                'amount_tax': 2 * atax,
-                'amount_total': 2 * (untax + atax),
+                'amount_tax': 3 * atax,
+                'amount_total': 3 * (untax + atax),
                 'amount_paid': 0,
                 'amount_return': 0,
             })
@@ -478,16 +495,21 @@ class TestPointOfSaleFlow(TestPointOfSaleCommon):
                 "active_id": pos_order.id,
             }
             pos_make_payment = self.PosMakePayment.with_context(context_make_payment).create({
-                'amount': 2 * (untax + atax),
+                'amount': 3 * (untax + atax),
             })
             context_payment = {'active_id': pos_order.id}
             pos_make_payment.with_context(context_payment).check()
 
             self.assertEqual(pos_order.state, 'paid')
-            self.assertEqual(pos_order.picking_ids.move_line_ids[0].lot_id, lot)
-            self.assertFalse(pos_order.picking_ids.move_line_ids[1].lot_id)
-            self.assertEqual(pos_order.picking_ids.move_line_ids[0].location_id, shelf1_location)
-            self.assertEqual(pos_order.picking_ids.move_line_ids[1].location_id, shelf1_location)
+            tracked_line_w_lot = pos_order.picking_ids.move_line_ids.filtered(lambda ml: ml.product_id.id == tracked_product_w_lot.id)
+            tracked_line_wo_lot = pos_order.picking_ids.move_line_ids.filtered(lambda ml: ml.product_id.id == tracked_product_wo_lot.id)
+            untracked_line = pos_order.picking_ids.move_line_ids - (tracked_line_w_lot | tracked_line_wo_lot)
+            self.assertEqual(tracked_line_w_lot.lot_id, lot)
+            self.assertTrue(tracked_line_wo_lot.lot_id)
+            self.assertFalse(untracked_line.lot_id)
+            self.assertEqual(tracked_line_w_lot.location_id, shelf1_location)
+            self.assertEqual(tracked_line_wo_lot.location_id, pos_order.config_id.picking_type_id.default_location_src_id)
+            self.assertEqual(untracked_line.location_id, shelf1_location)
 
         self.pos_config.current_session_id.action_pos_session_closing_control()