Rewrite `stock_picking._create_move_from_pos_order_lines`
to make calls in batches. Change the order of some operations
so that methods such as `stock.move._action_confirm()` and
`stock.production.lot.create` can be called only once.

Modify `test_order_to_pickings02` to add a check
on product tracked by lot but without a preexisting
`stock.production.lot`. i.e. check that
`_create_move_from_pos_order_lines` properly creates
a new lot.

#### speedup

A general speedup is quite difficult to establish as the time to create pickings at closing
of pos sessions depends on a lot of factors such as stock_account being installed,
orders with tracked products, orders with products with phantom-type boms, anglo-saxon
accounting etc..

Here, the times reported are about closing 1 session in a customer DB with
147 orders, 669 order_lines and 95 distinct products. Stock_account is installed,
anglo-saxon accounting is turned off. Most products have phantom-type boms with between
1 and 37 bom components. The products inventory valuation is set to real_time.

| distinct products | order lines | Before PR | After PR |
|:----------------------:|:--------------:|:--------------:|:-----------:|
| 10 | 11 | 6m | 5s |
| 25 | 33 | 17m | 7.88s |
| 50 | 107 | 30m | 10.98s |
| 95 | 637 | 35m | 15.44s|


##### NB for review: 
this PR should probably not be FP to 15.0 onwards as there were some changes
done to stock.picking in v15. I'll do another PR for v15 after this one gets merged.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr